ms-roles-api/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pytest==7.4.3
httpx==0.25.2
azure-cosmos==4.5.1
python-dotenv==1.0.0
